+++ src/js/render-movies.js
@@ -1,36 +1,110 @@
 import { refs } from './refs';
-import { renderGallery } from './templates/render-gallery';
+import pagination from './pagination';
+// import './paginationBtnWorks';
 import FilmsApiService from './fetch-api';
+import { renderGallery } from './templates/render-gallery';
+import './render-movies';
 
 const filmsApi = new FilmsApiService();
 
+popular();
 
-filmsApi.fetchTrendingFilms().then(response => {
-  renderGallery(response.results);
-  
-});
+function popular() {
+  filmsApi.fetchTrendingFilms().then(r => {
+    const page = r.page;
+    const totalPages = r.total_pages;
+    renderGallery(r.results);
+    pagination(totalPages, page);
+  });
+}
 
+function search() {
+  filmsApi.fetchFilmsOnSearch().then(r => {
+    const page = r.page;
+    const totalPages = r.total_pages;
+    renderGallery(r.results);
+    pagination(totalPages, page);
+  });
+}
+
+const ulTag = document.querySelector('.pagination ul');
+
+ulTag.addEventListener('click', onClick);
+
+function onClick(evt) {
+    if (evt.target.nodeName !== 'LI') {
+        return;
+    }
+    console.log(evt.target);
+    refs.gallery.innerHTML = '';
+    const allNumbes = document.querySelectorAll('.numb');
+    const arrowRight = document.querySelector('.arrow-right');
+    const arrowLeft = document.querySelector('.arrow-left');
+    const svgRight = document.querySelector('.arrow-right svg');
+    const svgPathRight = document.querySelector('.arrow-right svg path');
+    const svgLeft = document.querySelector('.arrow-left svg');
+    const svgPathLeft = document.querySelector('.arrow-left svg path');
+
+    if ( 
+        evt.target === arrowLeft ||
+        evt.target === svgLeft ||
+        evt.target === svgPathLeft
+      ) {
+        console.log('Вы нажали кнопку влево');
+        filmsApi.pageDecrement();
+        if (filmsApi.searchQuery !== null && filmsApi.searchQuery !== '') {
+          search();
+        } else popular();
+        // console.log(Number(document.querySelector('.numb').textContent));
+    }
+
+    // const allNumbes = document.querySelectorAll('.numb');
+    for (const number of allNumbes) {
+        if (evt.target === number) {
+            console.log('Вы нажали кнопку с цифрой');
+            filmsApi.setPage(Number(evt.target.textContent));
+            // filmsApi.pageDecrement();
+            if (filmsApi.searchQuery !== null && filmsApi.searchQuery !== '') {
+              search();
+            } else popular();
+        }
+    }
+
+    if (
+        evt.target === arrowRight ||
+        evt.target === svgRight ||
+        evt.target === svgPathRight
+    ) {
+        console.log('Вы нажали кнопку вправо');  
+        filmsApi.pageIncrement();
+        // filmsApi.pageDecrement();
+        if (filmsApi.searchQuery !== null && filmsApi.searchQuery !== '') {
+          search();
+        } else popular();
+    }
+}
 
 refs.searchForm.addEventListener('submit', onSearchSubmit);
 
 function onSearchSubmit(event) {
-  event.preventDefault();
+    event.preventDefault();
+    filmsApi.resetPage();
+     
+    filmsApi.query = event.target.elements.userSearchQuery.value.trim();
 
-  filmsApi.query = event.target.elements.userSearchQuery.value.trim();
+        if (filmsApi.searchQuery === "" || filmsApi.searchQuery === " ") {
+          refs.searchError.classList.remove('is-hidden');
+          return;
+        };
 
-    if (filmsApi.searchQuery === "" || filmsApi.searchQuery === " ") {
-      refs.searchError.classList.remove('is-hidden');
-      return;
-    };
-
-  filmsApi.fetchFilmsOnSearch().then(response => {
-     if (response.total_results === 0) {
-      refs.searchError.classList.remove('is-hidden');
-      return;
-    };
+    filmsApi.fetchFilmsOnSearch().then(response => {
+        if (response.total_results === 0) {
+          refs.searchError.classList.remove('is-hidden');
+          return;
+        };
 
     refs.searchError.classList.add('is-hidden'); 
     renderGallery(response.results);
-    } 
-  );
+    ulTag.addEventListener('click', onClick);
+    });
 };
